remove contractUtils file since only some methods in use
